% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{alternative_parents}
\alias{alternative_parents}
\alias{alternative_parents,ClassyFire-method}
\title{alternative_parents}
\usage{
alternative_parents(object)

\S4method{alternative_parents}{ClassyFire}(object)
}
\arguments{
\item{object}{a \code{ClassyFire} S4 object}
}
\description{
Get the alternative_parents
}
