% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{chebi}
\alias{chebi}
\alias{chebi,ClassyFire-method}
\title{chebi}
\usage{
chebi(object)

\S4method{chebi}{ClassyFire}(object)
}
\arguments{
\item{object}{a \code{ClassyFire} S4 object}
}
\description{
Get the predicted ChEBI identifications
}
