% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ClassyFire-class}
\alias{ClassyFire-class}
\title{ClassyFire Class}
\description{
A S4 class to store the results from the ClassFire RESTful API
}
\section{Slots}{

\describe{
\item{\code{meta}}{a list of the input SMILE, InChI Key and classification version used}

\item{\code{classification}}{a \code{tibble} of the ClassyFire classification results}

\item{\code{direct_parent}}{a list of the direct parent}

\item{\code{alternative_parents}}{a \code{tibble} of alternative parents}

\item{\code{predicted_chebi}}{a character vector of the predicted ChEBI terms}

\item{\code{external_descriptors}}{a \code{tibble} of any external identificatons}

\item{\code{description}}{a character string of the compound description}
}}

