% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{description}
\alias{description}
\alias{description,ClassyFire-method}
\title{description}
\usage{
description(object)

\S4method{description}{ClassyFire}(object)
}
\arguments{
\item{object}{a \code{ClassyFire} S4 object}
}
\description{
Get the \code{ClassyFire} description
}
