% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{descriptors}
\alias{descriptors}
\alias{descriptors,ClassyFire-method}
\alias{descriptors,Query-method}
\title{descriptors}
\usage{
descriptors(object)

\S4method{descriptors}{ClassyFire}(object)

\S4method{descriptors}{Query}(object)
}
\arguments{
\item{object}{a \code{Query} S4 object}
}
\value{
a \code{tbl_df} of available external descriptors

a \code{tbl_df} of available external descriptors
}
\description{
Get the external descriptors

Get the external descriptors
}
