% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_classification.R
\name{entity_classification}
\alias{entity_classification}
\title{Entity Classification}
\usage{
entity_classification(INCHI_KEY)
}
\arguments{
\item{INCHI_KEY}{a character string of a valid InChIKey}
}
\value{
a list of two elements
  \itemize{
      \item{ClassyFire}
      \item{Meta}

  }
}
\description{
Entity Classification
}
\examples{
entity_classification('BRMWTNUJHUMWMS-LURJTMIESA-N')

}
