% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status_code.R
\name{get_status_code}
\alias{get_status_code}
\title{Get Status Code}
\usage{
get_status_code(query_id)
}
\arguments{
\item{query_id}{a numeric value for the query id}
}
\value{
a list of \code{query_id} and classification status; either \code{In progress} or \code{Done}
}
\description{
Retrieve the status code of a new submission for entity classification
}
\examples{
get_status_code(2813259)
}
