% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{meta}
\alias{meta}
\alias{meta,ClassyFire-method}
\title{meta}
\usage{
meta(object)

\S4method{meta}{ClassyFire}(object)
}
\arguments{
\item{object}{a \code{ClassyFire} S4 object}
}
\description{
Get the \code{ClassyFire} meta data
}
