% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification.R
\name{open_cache}
\alias{open_cache}
\title{Open Cache}
\usage{
open_cache(dbname = ":memory:")
}
\arguments{
\item{dbname}{The path to the database file. SQLite keeps each database instance
in one single file. The name of the database is the file name, thus database names
should be legal file names in the running platform. There are two exceptions:
"" will create a temporary on-disk database. The file will be deleted when the connection is closed.
":memory:" or "file::memory:" will create a temporary in-memory database.}
}
\value{
\code{conn} a DBIConnection object, as produced by dbConnect
}
\description{
Creates a SQLight database for the local caching. This database includes already
queried Inchikeys and the serialized classification object.
}
