% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{parse_external_desc}
\alias{parse_external_desc}
\title{Parse External Descriptors}
\usage{
parse_external_desc(x)
}
\arguments{
\item{json_res}{the list output from \code{jsonlite::fromJSON(x)}}
}
\value{
a \code{tibble} containing the following;
\itemize{
\item \strong{source} External source name
\item \strong{source_id} External source ID
\item \strong{annotations} External source annotation
}
}
\description{
Parse the list output of returned External Descriptors
}
\keyword{internal}
