% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{parse_json_output}
\alias{parse_json_output}
\title{Parse JSON output}
\usage{
parse_json_output(json_res)
}
\arguments{
\item{json_res}{the list output from \code{jsonlite::fromJSON(x)}}
}
\value{
a \code{tibble} containing the following;
- \strong{Level} Classification level (kingdom, superclass, class and subclass)
- \strong{Classification} The compound classification
- \strong{CHEMONT} Chemical Ontology Identification code
}
\description{
Parse the JSON output the the \code{GET} or \code{POST} request
}
\keyword{internal}
