% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_classification.R
\name{retrieve_classification}
\alias{retrieve_classification}
\title{Retrieve classification results}
\usage{
retrieve_classification(query_id)
}
\arguments{
\item{query_id}{a numeric value for the query id}
}
\value{
a \code{tibble} containing the following;
\itemize{
\item \strong{Level} Classification level (kingdom, superclass, class and subclass)
\item \strong{Classification} The compound classification
\item \strong{CHEMONT} Chemical Ontology Identification code
}
}
\description{
Retrieve classification results from a new submission
}
\examples{

retrieve_classification(2813259)
}
