% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\name{show,ClassyFire-method}
\alias{show,ClassyFire-method}
\title{show-ClassyFire}
\usage{
\S4method{show}{ClassyFire}(object)
}
\arguments{
\item{object}{a \code{ClassyFire} S4 object}
}
\description{
show-ClassyFire
}
