% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{submit_query}
\alias{submit_query}
\title{Submits a ClassyFire query in a JSON format.}
\usage{
submit_query(label, input, type = "STRUCTURE")
}
\arguments{
\item{label}{\link{String} the label of the query.}

\item{input}{A named list or vector of SMILES strings.}

\item{type}{\link{String} the type of the query.}
}
\value{
A list of tibbles named by input name.
}
\description{
Submits a ClassyFire query in a JSON format.
}
\examples{
input <- c(MOL1 = 'CCCOCC', MOL2 = 'COCC=CCC')
submit_query(label = 'query_test', input = input, type = 'STRUCTURE')

}
