% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{unclassified}
\alias{unclassified}
\alias{unclassified,Query-method}
\title{unclassified}
\usage{
unclassified(object)

\S4method{unclassified}{Query}(object)
}
\arguments{
\item{object}{a \code{Query} S4 object}
}
\value{
a character string of \code{inputs} that were not successfully classified
}
\description{
Get the identifiers of any inputs that were not classified
}
