% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{chebi}
\alias{chebi}
\alias{chebi,ClassyFire-method}
\alias{chebi,Query-method}
\title{chebi}
\usage{
chebi(object)

\S4method{chebi}{ClassyFire}(object)

\S4method{chebi}{Query}(object)
}
\arguments{
\item{object}{a \code{Query} S4 object}
}
\value{
a character vector of ChEBI terms

a character vector of ChEBI terms
}
\description{
Get the predicted ChEBI identifications

Get the predicted ChEBI identifications
}
