% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{classification}
\alias{classification}
\alias{classification,ClassyFire-method}
\alias{classification,Query-method}
\title{Accessor Methods for the ClassyFire S4 class}
\usage{
classification(object)

\S4method{classification}{ClassyFire}(object)

\S4method{classification}{Query}(object)
}
\arguments{
\item{object}{a \code{Query} S4 object}
}
\description{
Get the \code{ClassyFire} classification results

Get the \code{ClassyFire} classification results
}
