% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accessor-methods.R
\name{descriptors}
\alias{descriptors}
\alias{descriptors,ClassyFire-method}
\title{descriptors}
\usage{
descriptors(object)

\S4method{descriptors}{ClassyFire}(object)
}
\arguments{
\item{object}{a \code{ClassyFire} S4 object}
}
\description{
Get the external descriptors
}
