% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_classification.R
\name{entity_classification}
\alias{entity_classification}
\title{Entity Classification}
\usage{
entity_classification(INCHI_KEY)
}
\arguments{
\item{INCHI_KEY}{a character string of a valid InChIKey}
}
\value{
a list of two \code{tibbles}
\itemize{
\item \strong{ClassyFire} a \code{tibble} containing the following;
\itemize{
\item \strong{Level}
\item \strong{Classification}
\item \strong{CHEMONT}
}
\item \strong{Meta} a \code{tibble} containing the following;
\itemize{
\item \strong{Query}
\item \strong{Version}
\item \strong{Date}
}
}
}
\description{
Entity Classification
}
\examples{
entity_classification('BRMWTNUJHUMWMS-LURJTMIESA-N')

}
