% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_classification.R
\name{entity_classification}
\alias{entity_classification}
\title{Entity Classification}
\usage{
entity_classification(inchi_key)
}
\arguments{
\item{inchi_key}{a character string of a valid InChIKey}
}
\value{
a \code{tibble} containing the following;
- \strong{Level} Classification level (kingdom, superclass, class and subclass)
- \strong{Classification} The compound classification
- \strong{CHEMONT} Chemical Ontology Identification code
}
\description{
Retrive entity classification from `http://classyfire.wishartlab.com/entities/'
}
\examples{
entity_classification('BRMWTNUJHUMWMS-LURJTMIESA-N')

entity_classification('MDHYEMXUFSJLGV-UHFFFAOYSA-N')

}
