% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_classification.R
\name{entity_classification}
\alias{entity_classification}
\title{Entity Classification}
\usage{
entity_classification(inchi_key)
}
\arguments{
\item{inchi_key}{a character string of a valid InChIKey}
}
\value{
a list of two \code{tibbles}
\itemize{
\item \strong{ClassyFire} a \code{tibble} containing the following;
\itemize{
\item \strong{Level} Classification level (kingdom, superclass, class and subclass)
\item \strong{Classification} The compound classification
\item \strong{CHEMONT} Chemical Ontology Identification code
}
\item \strong{Meta} a \code{tibble} containing the following;
\itemize{
\item \strong{Query} the submitted InChI key
\item \strong{Version} Version of classyFire
\item \strong{Date} Date classification retrieved
}
}
}
\description{
Retrive entity classification from `http://classyfire.wishartlab.com/entities/'
}
\examples{
entity_classification('BRMWTNUJHUMWMS-LURJTMIESA-N')

entity_classification('MDHYEMXUFSJLGV-UHFFFAOYSA-N')

}
