% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification.R
\name{get_classification}
\alias{get_classification}
\title{Get Entity Classification}
\usage{
get_classification(inchi_key)
}
\arguments{
\item{inchi_key}{a character string of a valid InChIKey}
}
\value{
a \code{ClassyFire} S4 object.
}
\description{
Retrieve entity classification from `http://classyfire.wishartlab.com/entities/'
}
\examples{
\dontrun{

# Valid InChI key where all four classification levels are available
get_classification('BRMWTNUJHUMWMS-LURJTMIESA-N')

# Valid InChI key where only three classification levels are available
get_classification('MDHYEMXUFSJLGV-UHFFFAOYSA-N')

# Invalid InChI key
get_classification('MDHYEMXUFSJLGV-UHFFFAOYSA-B')
}
}
\seealso{
ClassyFire-class
}
