% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_query}
\alias{get_query}
\title{Retrieves the classification results for a given query.}
\usage{
get_query(query_id, format = c("json", "sdf", "csv"))
}
\arguments{
\item{query_id}{query_id a numeric value for the ID of the query.}

\item{format}{a string of the format of the query (either JSON, CSV, or SDF)}
}
\value{
the parsed text output displaying the classification results for
}
\description{
Retrieves the classification results for a given query.
}
