% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_classification.R
\name{submit_classification}
\alias{submit_classification}
\title{Submit InChI Code for Classification}
\usage{
submit_classification(query, label, type = "STRUCTURE")
}
\arguments{
\item{query}{a character string of \code{InChI Code} or \code{SMILE}}

\item{label}{a character string of the query name}

\item{type}{the label type (\code{Default = STRUCTURE})}
}
\value{
if the classification has completed;  a \code{tibble} containing the following;
\itemize{
\item \strong{Level} Classification level (kingdom, superclass, class and subclass)
\item \strong{Classification} The compound classification
\item \strong{CHEMONT} Chemical Ontology Identification code
}
}
\description{
Submit and new entity for classification using the ClassyFire webserver
}
